backend/requirements.txt
Flask
flask-cors
pandas
numpy
scipy
python-dotenv
google-genai
PyJWT
